o New file dmfs-super.c  o dmfs-dir.c becomes dmfs-lv.c  o dmfs-file.c becomes dmfs-table.c  o A few tweeks and updates
The main reason for the slow progress on these files (which are not yet used
by the device mapper) is that we are working out what this interface should
look like as we go along.

Once this has evolved a bit further and in a state where it can be used we'll
announce it on the lists for further comment.