o Confusingly, dmfs-tdir isn't gone, its now called dmfs-lv.c and its the    old dmfs-lv.c thats gone.  o Dropped out support for multiple tables in line with ioctl interface  o Some reordering to better support the userland library  o Updated to 2.4.16
I'm fairly happy with the way that this is working now, so the next job is
to start the integration with the ioctl interface so there is a single
common dm.[ch] and selectable interfaces (fs or ioctl).

Further improvements can be made even now, but I hope to wait until we've
got this going and integrated and the libdm parts working as well before
investigating other avenues.